[joystick] Wii: introduce split-joysticks option
Add an option to present the Wii controller expansions as independent
controllers.

Introduce the SDL_WII_JOYSTICK_SPLIT environment variable, which, if set
to "1" before initializing the joystick subsystem, changes the number of
reported joysticks from 8 to 12:

- Joysticks 0-3: the 4 wiimotes
- Joysticks 4-7: the 4 expansions
- Joysticks 8-11: the 4 GameCube controllers

This is for the benefits of those applications which need to alter their
behaviour depending on the controller that generated the event, or even
to increase the number of possible players in a game, assuming that it
needs just the directional keys and a couple of buttons.

Due to the refactoring, some changes were made to the default non-split
view of the controllers, a couple of which are actually bugfixes:

- When a classic controller is being used, the arrow keys (hat) on the
  wiimote also work, delivering the same events as the arrow keys on the
  classic controller (previously, pressing them would produce no effect
  whatsoever, which looks like a bug).

- When an expansion is detached, its axes are reset to 0; previously,
  they would continue to be reported as moving, therefore rendering the
  controller unusable.

- Naming of the Wii controllers has been extended and now looks as
  following (assuming that we are operating on the second wiimote),
  depending on the expansion attached:
  * `Wiimote 1` (if no expansion is attached)
  * `Wiimote 1 + Nunchuck`
  * `Wiimote 1 + Classic`
  * `Wiimote 1 + Guitar Hero 3`
  * `Wiimote 1 + Balance board`
  This can help applications to map buttons and axes to their proper
  function.