ogc: implement framebuffer access in video driver
Accessing the EFB in this way is extremely slow, because every pixel
access goes through a conversion pipeline. Yet, it's the easiest and
quickest way to get something shown on the screen, just to get the tests
running. We'll later replace it with an accelerated solution.

This works on the emulator only. On a real wii, we get a hard lockup if
writing to the uncached EFB, and display corruption if writing (like we
do here) to the cached EFB. Adding a call to DCStoreCache on the EFB
memory also causes a lockup.

In other words, this is not really working.