ogc: implement rendering of mouse cursor
The cursor is drawn (if it needs to) at the end of each frame, right
before copying the changes to the EFB.

The small change in SDL_render_ogc.c is needed because we need a
specific blending mode to draw the cursor, and the renderer will need to
reset it.

The SDL_ogccursors.h file contains bitmaps of cursors to use on the Wii.
The hand drawing is taken from
https://wiibrew.org/wiki/Wii_Homebrew_Cursors (all cursors there are
under public domain), the arrow cursor has been created by me in
inkscape. The files then have been exported in GIMP as C source file and
their data pasted into this file.