ogc: do not create a new mouse cursor at every frame
The system default cursor is not supposed to change, so we can create it
just once (when the mouse gets initialized). In the polling callback we
then just need to hide or show it.

The logic of showing is not changed: if the application has set a
specific cursor, then it's in full control of its visibility; but if
it's using the default cursor, then we show the cursor only if at least
one wiimote is pointed at the screen.