ogc: apply viewport scale in QueueGeometry
The `scale_x` and `scale_y` parameters to QueueGeometry need to be
applied to the geometry coordinates. Without this, the
SDL_RenderSetLogicalSize() would not properly work on the renderer.

Note that a similar scale does not need to be applied by the backend in
the other draw functions (line, fill, copyrect), because the coordinates
are already scaled then.