ogc: always give focus to the window
We support only one window and that should receive the focus. Failing to
do this breaks the delivering of Joystick and GameController events,
unless the SDL_JOYSTICK_ALLOW_BACKGROUND_EVENTS hints is set (which is
not, by default).

Some backends do this in CreateWindow(), some in ShowWindow(). In our
case it doesn't really make a difference, but ShowWindow() sounds more
correct.