ogc: add support APIs for virtual keyboards
This commit adds a new public header (SDL_ogcsupport.h) which provides a
couple of APIs so that an application would be able to hook into the
On-Screen Keyboard (OSK) calls and provide its own implementation of an
OSK.

The usefulness of this is that the API is designed in such a way that
the OSK implementation can reside in a separate project which would only
provide a static library to be linked against the application. In this
way, SDL applications only need change a couple of lines in order to get
the OSK functionality.