ogc: always let application decide on mouse visibility (#59)
Our heuristics break the "Abbaye des Morts" game, which does not need a
mouse cursor and just calls `SDL_ShowCursor(SDL_DISABLE)` at startup.
Yet, with the current version of SDL, if the wiimote is pointed at the
screen we are still showing a cursor, overriding the application's
explicit decision. And the mouse cursor is useless, since the
application is not reacting to a mouse.

The code we are removing was mainly added with the intent of _hiding_
the mouse cursor when the wiimote was not pointed at the screen, and we
could indeed update it and keep it working just as an additional check
on whether the cursor should be shown or not, but we don't really have a
concrete use-case here.

Let's just remove it completely for now, and bring it back (still
respecting an explicit SDL_DISABLE setting, though!) if/when we see a
reason to.