ogc: maintain scale of mouse cursor
The mouse cursor need to get scaled accordingly to the screen
resolution: it was designed for a resolution of 640x480, but on video
modes having a different aspect ratio (such as 640x240) or a different
scale (such as 320x240) it will appear deformed or huge.

Use a model view matrix to transform it appropriately.