ogc: flush the FIFO after executing the command queue
It can happens that the client program immediately calls
SDL_DestroyTexture() after draawing it. Luckily for us, SDL is smart
enough to realize that the texture has been used in a draw operation and
issues a SDL_RenderFlush() so that all the operations get executed. But
this means that on our side, we must also make sure that when we return
from OGC_RunCommandQueue() the texture data is no longer needed. The
call to GX_DrawDone() does exactly this.