ogc: save and restore EFB when rendering to a texture starts and stops
Indeed, the situation where a rendering to texture operation begins
after the EFB has already been painted can happen. In order not to lose
the existing EFB contents, make a copy of them into a texture (using the
same size of the render-to-texture target) and then restore them by
redrawing the saved texture over the EFB, once the render-to-texture
operation has completed.

Note that it's important that any calls to GX_SetPixelFmt() happen in
the appropriate order:

1. Save the current EFB contents
2. Change the EFB format (by calling GX_SetPixelFmt())
3. Perform all the render-to-texture drawing operations
4. Save the EFB into the target texture
5. Restore the EFB format
6. Restore the EFB contents by drawing the texture saved in step 1

Swapping the 5th and 6th steps does not cause any visible issues in
Dolphin, but causes the EFB to be painted back with wrong colors on a
real Wii.